Make AbstractValues inner class of AbstractOptions
This reduces boilerplate because derived classes don't have to define a
constructor for the AbstractValues subclass.
