+++ src/main/java/org/scijava/optional/AbstractOptions.java
@@ -2,6 +2,7 @@ package org.scijava.optional;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.function.BiConsumer;
 
 public abstract class AbstractOptions< T extends AbstractOptions< T > > implements Options< T >
 {
@@ -9,7 +10,7 @@ public abstract class AbstractOptions< T extends AbstractOptions< T > > implemen
 
 	protected AbstractOptions( T that )
 	{
-		that.theOptions.forEach( theOptions::put );
+		theOptions.putAll( that.theOptions );
 	}
 
 	public AbstractOptions()
@@ -27,7 +28,11 @@ public abstract class AbstractOptions< T extends AbstractOptions< T > > implemen
 			return ( T ) this;
 
 		T concat = copyOrThis();
-		additionalOptions.theOptions.forEach( concat.theOptions::put );
+		additionalOptions.theOptions.forEach( ( k, v ) -> {
+			// NB remove existing key for ordering for appended options
+			concat.theOptions.remove( k );
+			concat.theOptions.put( k, v );
+		} );
 		return concat;
 	}
 
@@ -35,6 +40,7 @@ public abstract class AbstractOptions< T extends AbstractOptions< T > > implemen
 	public T add( final String key, final Object value )
 	{
 		final T copy = copyOrThis();
+		// NB remove existing key for ordering for appended options
 		copy.theOptions.remove( key );
 		copy.theOptions.put( key, value );
 		return copy;
@@ -56,5 +62,63 @@ public abstract class AbstractOptions< T extends AbstractOptions< T > > implemen
 		}
 		sb.append( "}" );
 		return sb.toString();
+	}
+
+	protected abstract class AbstractValues implements Values
+	{
+		public void forEach( BiConsumer< String, Object > action )
+		{
+			theOptions.forEach( action );
+		}
+
+		@Override
+		public String toString()
+		{
+			final ValuesToString sb = new ValuesToString();
+			forEach( sb );
+			return sb.toString();
+		}
+
+		@Override
+		public < T > T value( final String key, final T defaultValue )
+		{
+			@SuppressWarnings( "unchecked" )
+			final T value = ( T ) theOptions.get( key );
+			return value == null ? defaultValue : value;
+		}
+	}
+
+	protected class ValuesToString implements BiConsumer< String, Object >
+	{
+		private final StringBuilder sb;
+
+		private boolean first;
+
+		public ValuesToString()
+		{
+			sb = new StringBuilder().append( "{" );
+			first = true;
+		}
+
+		@Override
+		public String toString()
+		{
+			sb.append( "}" );
+			return sb.toString();
+		}
+
+		@Override
+		public void accept( final String key, final Object value )
+		{
+			if ( first )
+				first = false;
+			else
+				sb.append( ", " );
+			sb.append( key );
+			sb.append( " = " );
+			sb.append( value );
+			if ( !theOptions.containsKey( key ) )
+				sb.append( " [default]" );
+		}
 	}
 }
